refactor: combat phase-2 slice 7 — QB perf and build/siege tables (#3865)
Migrate quick_battle_resolver_perf_invariants_test and
quick_battle_resolver_build_and_siege_test to table-driven scenario loops
with shared builders in colonizethis_combat_test_support; combat test LOC
4756 → 4156.

Refs #3865
